requirements.txt
pytest~=7.1.3
selene>=2.0.b8
allure-pytest
python-dotenv==0.21.0
Appium-Python-Client~=2.6.1
selenium~=4.2.0
requests
pydantic
